Add PO - Programming with Objects
This class was about OOP design patterns. The language was Java, with
some bonus if C++ exercises were made.

I only wrote down the C++ exercises, because most Java exercises came
with solutions.

Regarding the project, it was to make a "Make My Trip" (mmt) app.
Basically, we had to use a pre-existing library, po-uuilib.jar, and
some application code, not complete (mmt-app), and write the core
(mmt-core), that had the concepts of the domain: TrainCompany,
TicketOffice(façade), Passenger, etc.

This class has a online wiki:

https://www.l2f.inesc-id.pt/~david/w/pt/Programa%C3%A7%C3%A3o_com_Objectos

The end project note was 18/20. The discount came mostly from not
passing some automated tests, due to a missunderstanding of the
concepts, and mis-use of maps. (Over use).