Add SO - Operating Systems
Despite the name, this class was not only about OS's, but also about
concurrent and parallel programming, with both processes and threads,
using shared memory and messages, and interrupts.

The project was divided in 4 parts; but because each part built on
previous code, there is only 1 project folder. The parts are in
Question_papers.zip. The Makefile can build any version of the 4
delivered, just run

	make v[0-3]

The project was a heat simulation - we created a matrix of NxN size,
set the initial border temperatures, and then the next cell temperature
was calculated, by the average of the neighbours previous temperature.
The code to manipulate and create matrixes was given, as well as the
message passing library; we only had to write the argument "parsing" and
the calculus of the next temperature, as well as displaying the result,
and save the matrix while doing calculations.

As a bonus, because we (group of 2) were pretty thorough on checking
errors, and the code was hard to read, we added the
rem_error_checking.py, that is a crude script to try to remove most of
the checks, providing easier to read code (but probably won't compile).

The project note was 20/20, with no rounding up.

Out of respect to the privacy of my colleague, I won't include the repo
history.