Add ASA - Analysis and Synthesis of Algorithms
+ Update README to include 2nd year, 2nd semester
+ Update LICENSE to include 2018

In this class, we learned in more detail how to analyse algorithms,
algorithms on graphs, plus greedy vs dinamic programming algorithms,
linear programming and a more pratical view on NP problems (instead of
non-determinism, the time to generate vs time to verify a solution).

The exercises made on pratical classes were not programming, so no LAB
folder.

The first project was to find Strongly Connected Components (SCCs),
implementing either Tarjan or the DFS-based algorithm (or other, but we
did not learn more in classes). We used Tarjan.

The second project was classification of "pixels" in an image, based on
weigths in both each pixel and neighbourhood. The solution was to use a
min-cut / max-flow algorithm, like Ford-Fulkerson. We implemented the
Edmonds-Karp variant (using a BFS to search paths).

In addition to verify the implementation, we also had to write a report
with a description of the problem, our solution, and theoretical and
pratical big O measures of our implementation.

In both projects, the grade was 20/20.

This time around, the full repo with history is directly in the folders.

Changes to be committed:
	new file:   ASA/Proj1
	new file:   ASA/Proj2
	modified:   LICENSE
	modified:   README.md