Add IPM - Person-Machine Interfaces
In this class, we learned about how to create programs and solutions
that help users complete their tasks, and how to design them, as well as
how to use iterative development guided by user and expert testing,
using prototypes of low and high fidelity, and best practices.

The project was to make a high-fidelity prototype of a smartwatch for
summer music festivals - the iRave. The code part of the work was not
the most important, and was not used for grading -- only the interaction
mattered. Also, first-time learning and using JS, and jQuery -- this
project is a big ball of mud and anti-patterns.

In the end, this prototype averaged 19/20 over 3 functionalities.

Out of respect to the privacy of my colleagues, I won't include the repo
history. It isn't very interesting.

Changes to be committed:
	new file:   IPM/Proj.zip