Add IA - Artificial Inteligence
+ Update LICENSE to encompass 2019

As the name hints, this class was an overview of the field of Artificial
Inteligence, from agents to classification algorithms and rudimentary
"machine learning".

The first project was to compare 3 different search algorithms on the
problem of "Peg Solitaire". The 3 algorithms used were Depth-first,
Greedy and A* searches. The heuristic used was non-admissible, but
nevertheless we succeeded in solving the problems in reasonable time
(and overall the A* performed well in the most complex setups).

The second project was a mix of 2 problems. The first problem was to
implement bayesian net functions (with Python and numpy), and was
relatively trivial. The second problem was about reinforcement learning
and Q-Learning, and fidled with learning parameters (alpha and training
traject/time); also implemented some trivial functions for Q-Learning.

In the first project, we got 20/20; in the second, 19.5/20.