Add RC - Computer Networks
In this class, we learnt about the basics of networking, specially the
base protocols that make up the Internet as we know it.

The project was to implement a basic protocol and network application to
backup and restore user files.
The components are:
  - CS.py: A Central Server, acting as a coordinator
  - BS.py: The Backup Server, which must listen for backup & restore
    requests
  - user.py: The user "interface" application

The BSs and CS coordinated among them using UDP. All client
communications went over TCP.

In this project, our group got 19/20.